CMakeDetermineCompilerId: Fix compiler id with square brackets in the path
We use file(GLOB) to load the list of files produced by compiling the
compiler identification source.  Encode square brackets in the path to
the directory so that they are not treated as special characters in the
globbing expression.  Otherwise we fail to find any files when the path
contains square brackets and the compiler id is unknown.

Reported-by: Esch Nigma <eschnigma@openmailbox.org>