daemon returns errors when error occurs.
currently daemon returns nothing, which effectively hangs the client
if it permanently waits for daemon response. this adds error return
in the daemon when error occurs, thus unblocking the client. another
approach is to always have the client to use timed wait. but I think
a good server should try to be responsive in all cases.