Copyright.txt: Add notice of copyright by contributors
We do not require copyright assignment for contributions to CMake.
Update the top-level notice to mention "Contributors" explicitly.
Add a list of the Contributors whose names already appear in
source-level copyright notices.  Find the list with the command:

    $ git grep -h 'Copyright[^.]' -- *.* \
        Auxiliary Modules Packaging Source Templates Tests \
      | grep -v Kitware | sed 's/^#\? *//' | sort | uniq

Manually refine the list to de-duplicate entries and filter out
notices from third-party code.