Utilities/Release: Add html docs to Windows 64-bit binary
The 64-bit Windows binary is built in our second batch because it is
built on the same host as the 32-bit Windows binary.  When the
documentation generation was first added, it was not included in the
second batch because at the time that batch only included the Cygwin
binary.  Therefore the 64-bit Windows binary has been missing the
documentation since it was first created for CMake 3.6.

Since then the Cygwin binary was dropped so the second batch only
contains the 64-bit Windows binary.  Add the docs to it.

Closes: #16333