FindProtobuf: Fix protobuf_generate_*() to handle subdirs
The change in commit d4b8e81 incorrectly adds subdirectories to the
path of the generated files when *.proto files are passed to
protobuf_generate_*() with subdirectories. This behavior is not correct
when PROTOBUF_GENERATE_CPP_APPEND_PATH is True (default behavior) as
`protoc` will generate output file in the current build directory.